ragengine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Core RAG dependencies
pinecone-client>=3.1.0
google-generativeai>=0.4.0
# Document processing
pdfplumber>=0.10.3
python-docx>=1.1.0
python-magic>=0.4.27
tiktoken>=0.5.2
httpx>=0.25.2
pydantic>=2.5.0
email-validator>=2.1.0
requests>=2.31.0